readding _8

+++ src/ip2lib_8.h
@@ -185,7 +185,7 @@ void gdswzd(long igdtnum, long *igdtmpl, long igdtlen, long iopt,
  *
  * @author Jovic @date 2016
  */
-void gdswzd_grib1(long long int kgds, long long int iopt, long long int npts,
+void gdswzd_grib1(long kgds, long iopt, long npts,
                   double *fill, double *xpts,
 		  double *ypts, double *rlon, double *rlat, int nret, double *crot,
 		  double *srot, double *xlon, double *xlat, double *ylon,
